pyproject.toml
[tool.isort]
profile = "black"

[tool.poetry]
name = "ghostcompanion"
version = "0.1.0"
description = "Transfer transactions to Ghostfolio from numerous sources."
authors = ["Rafael Oliveira <olirafa@proton.me>"]
license = "The Unlicense"
readme = "README.md"

[tool.poetry.dependencies]
python = "^3.12"
requests = "^2.31.0"
python-dotenv = "^1.0.0"
pyyaml = "^6.0.2"
tastytrade = "^11.0.3"
yfinance = "^0.2.52"
pyjwt = "2.8.0"
cryptography = "42.0.5"
ibflex = {git = "https://github.com/agusalex/ibflex"}

[tool.poetry.group.dev.dependencies]
black = "^23.11.0"
isort = "^5.12.0"
pylama = "^8.4.1"
commitizen = "^3.12.0"
pytest = "^8.3.4"
setuptools = "^80.9.0"

[build-system]
requires = ["poetry-core"]
build-backend = "poetry.core.masonry.api"

[tool.pytest.ini_options]
addopts = "-vv"
testpaths = ["tests"]
pythonpath = ["."]
python_functions = "should_* when_*"
